Re-organize files for CG code
