+++ python/rascaline/rascaline/utils/clebsch_gordan/__init__.py
@@ -1,1 +1,1 @@
-from .correlate_density import DensityCorrelations  # noqa: F401
+from ._correlate_density import DensityCorrelations  # noqa: F401
